+++ source/Internal/Transition/Utility/BasicContext.php
@@ -203,4 +203,12 @@ class BasicContext implements BasicContextInterface
         }
         return $this->facts;
     }
+
+    public function getTemplateCacheDirectory(): string
+    {
+        return Path::join(
+            $this->getCacheDirectory(),
+            'template_cache'
+        );
+    }
 }