+++ tests/Unit/Internal/ContextStub.php
@@ -24,7 +24,7 @@ class ContextStub extends BasicContextStub implements ContextInterface
     private $isAdmin;
     private $skipLogTags;
     private $adminUserId;
-    private $productiveMode;
+    private bool $productiveMode;
 
     /**
      * ContextStub constructor.
@@ -41,6 +41,7 @@ class ContextStub extends BasicContextStub implements ContextInterface
         $this->doLogAdminQueries = $context->isEnabledAdminQueryLog();
         $this->isAdmin = $context->isAdmin();
         $this->skipLogTags = $context->getSkipLogTags();
+        $this->productiveMode = $context->isShopInProductiveMode();
     }
 
     /**
@@ -211,18 +212,12 @@ class ContextStub extends BasicContextStub implements ContextInterface
         $this->skipLogTags = $skipLogTags;
     }
 
-    /**
-     * @return bool
-     */
     public function isShopInProductiveMode(): bool
     {
         return $this->productiveMode;
     }
 
-    /**
-     * @param bool $productiveMode
-     */
-    public function setShopInProductiveMode(bool $productiveMode)
+    public function setShopInProductiveMode(bool $productiveMode): void
     {
         $this->productiveMode = $productiveMode;
     }